Use field coordinates for targetting
Instead of directly running a PID controller off the limelight,
use the data from the limelight to triangulate our position on
the field and compute a desired encoder value for the turret.

This commit changes the target tracker to essentially convert the
the distance and angle from the limelight to field coordinates and
then immediately back to a distance and angle again. It is not a
major functionality change; however, it will cause a couple subtle
behavior changes:

- When we don't have a limelight sample, the turret will aim
    at the last known location of the target, instead of wandering
    aimlessly until hitting the edge of its range. This is because
    the PID loop now tracks based on the encoders and NavX instead
    of raw camera values.
- Since we target using the encoders, the PID loop should no longer
    be affected as significantly by camera latency. This might help
    with the turret recoil problem, and we may additionally be able
    increase accuracy by tuning the PID loop tighter than what was
    previously possible.

Field-oriented targetting is a prerequisite for some of ou future
goals, including:

- Camera latency compensation
- Inner port targetting
- Shooting on the move